requirements.txt
streamlit>=1.26.0
requests
ipykernel
python-dotenv
openai
pydantic
shelve
